moving backfill to common step and other minor fixes
